backend/requirements.txt
fastapi
uvicorn
httpx[http2]
python-dotenv
beautifulsoup4
lxml
tldextract
dnspython
python-whois
